feat(Added e2e tests):
Added e2e tests. Fixed issue 3.
https://github.com/Dwctor/stm_add/issues/3
